css/css-backgrounds/parsing/border-image-outset-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing border-image-outset with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-image-outset">
<meta name="assert" content="border-image-outset supports the full grammar '[ <length> | <number> ]{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
// "1px" in Blink, Edge, WebKit; "1px 1px 1px 1px" in Firefox.
test_valid_value("border-image-outset", "1px", ["1px", "1px 1px 1px 1px"]);
// "2" in Blink, Edge, WebKit; "2 2 2 2" in Firefox.
test_valid_value("border-image-outset", "2", ["2", "2 2 2 2"]);

test_valid_value("border-image-outset", "1px 2 3px 4");
</script>
</body>
</html>
